Unify path handling in `FileManager.download` and `DKISTClient.fetch`
